LayoutTests/fast/js/dfg-arguments-alias-expected.txt

Tests aliased uses of 'arguments'.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS bar(42) is 42
PASS bar(42) is 42
PASS bar(42) is 42
PASS bar(42) is 42
PASS bar(42) is 42
PASS bar(42) is 42
PASS bar(42) is 42
PASS bar(42) is 42
PASS bar(42) is 42
PASS bar(42) is 42
PASS bar(42) is 42
PASS bar(42) is 42
PASS bar(42) is 42
PASS bar(42) is 42
PASS bar(42) is 42
PASS bar(42) is 42
PASS bar(42) is 42
PASS bar(42) is 42
PASS bar(42) is 42
PASS bar(42) is 42
PASS bar(42) is 42
PASS bar(42) is 42
PASS bar(42) is 42
PASS bar(42) is 42
PASS bar(42) is 42
PASS bar(42) is 42
PASS bar(42) is 42
PASS bar(42) is 42
PASS bar(42) is 42
PASS bar(42) is 42
PASS bar(42) is 42
PASS bar(42) is 42
PASS bar(42) is 42
PASS bar(42) is 42
PASS bar(42) is 42
PASS bar(42) is 42
PASS bar(42) is 42
PASS bar(42) is 42
PASS bar(42) is 42
PASS bar(42) is 42
PASS bar(42) is 42
PASS bar(42) is 42
PASS bar(42) is 42
PASS bar(42) is 42
PASS bar(42) is 42
PASS bar(42) is 42
PASS bar(42) is 42
PASS bar(42) is 42
PASS bar(42) is 42
PASS bar(42) is 42
PASS bar(42) is 42
PASS bar(42) is 42
PASS bar(42) is 42
PASS bar(42) is 42
PASS bar(42) is 42
PASS bar(42) is 42
PASS bar(42) is 42
PASS bar(42) is 42
PASS bar(42) is 42
PASS bar(42) is 42
PASS bar(42) is 42
PASS bar(42) is 42
PASS bar(42) is 42
PASS bar(42) is 42
PASS bar(42) is 42
PASS bar(42) is 42
PASS bar(42) is 42
PASS bar(42) is 42
PASS bar(42) is 42
PASS bar(42) is 42
PASS bar(42) is 42
PASS bar(42) is 42
PASS bar(42) is 42
PASS bar(42) is 42
PASS bar(42) is 42
PASS bar(42) is 42
PASS bar(42) is 42
PASS bar(42) is 42
PASS bar(42) is 42
PASS bar(42) is 42
PASS bar(42) is 42
PASS bar(42) is 42
PASS bar(42) is 42
PASS bar(42) is 42
PASS bar(42) is 42
PASS bar(42) is 42
PASS bar(42) is 42
PASS bar(42) is 42
PASS bar(42) is 42
PASS bar(42) is 42
PASS bar(42) is 42
PASS bar(42) is 42
PASS bar(42) is 42
PASS bar(42) is 42
PASS bar(42) is 42
PASS bar(42) is 42
PASS bar(42) is 42
PASS bar(42) is 42
PASS bar(42) is 42
PASS bar(42) is 42
PASS bar(42) is 42
PASS bar(42) is 42
PASS bar(42) is 42
PASS bar(42) is 42
PASS bar(42) is 42
PASS bar(42) is 42
PASS bar(42) is 42
PASS bar(42) is 42
PASS bar(42) is 42
PASS bar(42) is 42
PASS bar(42) is 42
PASS bar(42) is 42
PASS bar(42) is 42
PASS bar(42) is 42
PASS bar(42) is 42
PASS bar(42) is 42
PASS bar(42) is 42
PASS bar(42) is 42
PASS bar(42) is 42
PASS bar(42) is 42
PASS bar(42) is 42
PASS bar(42) is 42
PASS bar(42) is 42
PASS bar(42) is 42
PASS bar(42) is 42
PASS bar(42) is 42
PASS bar(42) is 42
PASS bar(42) is 42
PASS bar(42) is 42
PASS bar(42) is 42
PASS bar(42) is 42
PASS bar(42) is 42
PASS bar(42) is 42
PASS bar(42) is 42
PASS bar(42) is 42
PASS bar(42) is 42
PASS bar(42) is 42
PASS bar(42) is 42
PASS bar(42) is 42
PASS bar(42) is 42
PASS bar(42) is 42
PASS bar(42) is 42
PASS bar(42) is 42
PASS bar(42) is 42
PASS bar(42) is 42
PASS bar(42) is 42
PASS bar(42) is 42
PASS bar(42) is 42
PASS bar(42) is 42
PASS bar(42) is 42
PASS bar(42) is 42
PASS bar(42) is 42
PASS bar(42) is 42
PASS bar(42) is 42
PASS bar(42) is 42
PASS bar(42) is 42
PASS bar(42) is 42
PASS bar(42) is 42
PASS bar(42) is 42
PASS bar(42) is 42
PASS bar(42) is 42
PASS bar(42) is 42
PASS bar(42) is 42
PASS bar(42) is 42
PASS bar(42) is 42
PASS bar(42) is 42
PASS bar(42) is 42
PASS bar(42) is 42
PASS bar(42) is 42
PASS bar(42) is 42
PASS bar(42) is 42
PASS bar(42) is 42
PASS bar(42) is 42
PASS bar(42) is 42
PASS bar(42) is 42
PASS bar(42) is 42
PASS bar(42) is 42
PASS bar(42) is 42
PASS bar(42) is 42
PASS bar(42) is 42
PASS bar(42) is 42
PASS bar(42) is 42
PASS bar(42) is 42
PASS bar(42) is 42
PASS bar(42) is 42
PASS bar(42) is 42
PASS bar(42) is 42
PASS bar(42) is 42
PASS bar(42) is 42
PASS bar(42) is 42
PASS bar(42) is 42
PASS bar(42) is 42
PASS bar(42) is 42
PASS bar(42) is 42
PASS bar(42) is 42
PASS bar(42) is 42
PASS bar(42) is 42
PASS bar(42) is 42
PASS bar(42) is 42
PASS bar(42) is 42
PASS successfullyParsed is true

TEST COMPLETE
